Display message when snake is being validated
Prior to this commit, clicking on the Snake API Validation Info button
while the snake is being validated produced an error with the button
not responding at all.

This commit adds better user feedback to this button,
showing a `Snake is currently being validated ...` message while
the validation API call is still occurring.  If the snake API
validation call fails, it will also display a message.